… CHANGELOG.md
… CMakeLists.txt
… MERGE_PREPARATION.md
… README.md
… THREAD_PROGRESS.md
… src/common/version.h
… src/linux/v4l2/v4l2_capture.cpp